st/blit: blit-parameters extended to integers
Function glBlitFramebuffer takes parameters for blitting
as integers. Specification doesn't contain any restrictions
in minimum/maximum values of them so looks like they could
be any valid integer.

Gallium restricts some parameters to int16 (struct pipe_box).
Patch allows any integer to be used.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=108088

+++ src/gallium/include/pipe/p_state.h
@@ -517,10 +517,10 @@ struct pipe_box
     * x and width are used by buffers, so they need the full 32-bit range.
     */
    int x;
-   int16_t y;
+   int y;
    int16_t z;
    int width;
-   int16_t height;
+   int height;
    int16_t depth;
 };
